Healthcheck: split mongo test into separate handler.
Separating this from the SQL DB tests makes things easier to follow.
Keeping all these tests in a single handler like this would have only
increased in complexity as tests for more serices were added.
